Conflicts:
	src/server/scripts/EasternKingdoms/zone_duskwood.cpp

+++ src/server/game/Handlers/TradeHandler.cpp
@@ -299,14 +299,14 @@ void WorldSession::HandleAcceptTradeOpcode(WorldPacket& /*recvPacket*/)
         return;
     }
 
-    if (_player->GetMoney() >= uint32(MAX_MONEY_AMOUNT) - his_trade->GetMoney())
+    if (_player->GetMoney() >= MAX_MONEY_AMOUNT - his_trade->GetMoney())
     {
         _player->SendEquipError(EQUIP_ERR_TOO_MUCH_GOLD, NULL, NULL);
         my_trade->SetAccepted(false, true);
         return;
     }
 
-    if (trader->GetMoney() >= uint32(MAX_MONEY_AMOUNT) - my_trade->GetMoney())
+    if (trader->GetMoney() >= MAX_MONEY_AMOUNT - my_trade->GetMoney())
     {
         trader->SendEquipError(EQUIP_ERR_TOO_MUCH_GOLD, NULL, NULL);
         his_trade->SetAccepted(false, true);
@@ -686,7 +686,6 @@ void WorldSession::HandleSetTradeGoldOpcode(WorldPacket& recvPacket)
     if (!my_trade)
         return;
 
-    // gold can be incorrect, but this is checked at trade finished.
     my_trade->SetMoney(gold);
 }
